[TASK] Refactor facet ViewHelpers to use shared trait
- Extract shared methods into new `ValueViewHelperArgumentTrait` for reuse.
- Replace `AbstractValueViewHelper` inheritance with trait in `AddFacetItemViewHelper`, `SetFacetItemViewHelper`, and `RemoveFacetItemViewHelper`.
- Remove obsolete `AbstractValueViewHelper` class.
